Converted Tile.SIZE to float

+++ src/main/java/entity/dynamic/mob/Mob.java
@@ -182,8 +182,8 @@ public abstract class Mob extends Entity {
 
 	// DO NOT TOUCH THIS. SET THE MOVEMENT TO THE PATH OBJ USE move()!! DO NOT USE!!!
 	protected void moveTo(int x, int y) {
-		int xmov = Integer.compare(x*Tile.SIZE, (int) this.location.x);
-		int ymov = Integer.compare(y*Tile.SIZE, (int) this.location.y);
+		int xmov = Integer.compare((int) (x*Tile.SIZE), (int) this.location.x);
+		int ymov = Integer.compare((int) (y*Tile.SIZE), (int) this.location.y);
 		move(xmov*3, ymov*3);
 	}
