+++ src/main/java/entity/dynamic/mob/Zombie.java
@@ -8,7 +8,7 @@ public class Zombie extends Humanoid {
 
 	private byte animationtimer = 0;
 
-	public Zombie(Level[] levels, int x, int y, int z) {
+	public Zombie(Level[] levels, float x, float y, int z) {
 		super(levels, x, y, z);
 		sprite = SpriteHashtable.get(51);
 		setName("zombie");
